Use NCBI mapping data from older build
We previously used the NCBI mapping data for the current build, but now a new
build was just released, the mapping data is missing (they are waiting for the
data and it should be up soon).

For now, we just use the mapping data for the archived previous build (it is
the build we were using until now anyway).


git-svn-id: https://humgenprojects.lumc.nl/svn/mutalyzer/branches/refactor-mutalyzer-branch@361 eb6bd6ab-9ccd-42b9-aceb-e2899b4a52f1