Fix mapping info for genes mapped to more than one chromosome
Some genes (e.g. in the PAR) are mapped on both the X and Y chromosomes, but
are (apart from the chromosome names) indistinguishable from transcripts that
are mapped using different contigs. Transcripts of the latter type should be
merged, those of the former type should not be merged.

Our fix consists of only including exons where positions are consistent with
the transcript mapping and allowing transcripts to be mapped more than once,
but only to two different chromosomes.

This fixes #82.


git-svn-id: https://humgenprojects.lumc.nl/svn/mutalyzer/trunk@467 eb6bd6ab-9ccd-42b9-aceb-e2899b4a52f1