Refactor name checker to accept GET requests
Until now, the name checker form used POST requests. As a special case for
linking from LOVD, GET requests were handled by showing results without the
full interface (header, menu, etc). Ordinary linking was done to a separate
checkForward location which then set a cookie with the variant name and a
redirect to the name checker results. This was all very hackish and somewhat
broken (see Track issue #94).

This commit refactors the name checker to use GET requests, so ordinary
bookmarking of result pages is possible. All old entrypoints are handled with
a redirect for backwards compatibility.

Worth mentioning is that variant descriptions in the name checker are now
limited in length by the maximum query string length. However, this maximum
is several thousand characters with Internet Explorer having the lowest
maximum of just over 2000. Longer descriptions are not practically checked
with the name checker web interface anyway, so this should not be a problem.


git-svn-id: https://humgenprojects.lumc.nl/svn/mutalyzer/trunk@521 eb6bd6ab-9ccd-42b9-aceb-e2899b4a52f1