Port Mapping database module to SQLAlchemy
This introduces a proper notion of genome assemblies. Transcript
mappings for alle genome assemblies are in the same database, which
is better for maintenance. Updating transcript mappings is also
simplified a lot, especially from NCBI mapview files where we now
require a preprocessing sort on the input file.

Overall, this port touches a lot of Mutalyzer code, so beware.