Always give absolute template_folder and static_folder to Flask
For some reason, `pkg_resources.resource_filename` returns a relative path
if we call Mutalyzer using `python -m mutalyzer.entrypoints.*`. The path
is of course valid, but we pass it to Flask, which assumes relative paths
to be relative to the Python package root. To circumvent this incorrect
assumption, we first make the path absolute before passing it to Flask.