Don't auto remove comma characters in syntax checker
The web interface for the syntax checker previously automatically
removed comma characters from the description before parsing it. This
is not a good idea, mainly because comma characters are actually
allowed by the nomenclature (to distinguish different molecules on
RNA level).

Furthermore, this behaviour was a special case for the syntax checker
website interface and not present in either the batch or webservice
interfaces, nor in the name checker or any other interface.