In LRG, only support one CDS per transcript (ignore all others)
The LRG model allows for more than one CDS per transcript (and each CDS
can be coupled with more than one transcript). The Mutalyzer gene model
does not allow that.

Previously, multiple CDSs per transcript where treated incorrectly (I
think), so now we just ignore any CDS except the first per transcript.

Unfortunately, it is a bit more involved to give good warnings for this,
so that is not implemented.