Fix POST requests to the HTTP/RPC+JSON webservice
Upstream Spyne crashes on POST requests to the HTTP/RPC+JSON webservice.
We patched it in a rather hacky way.

This was a regression from the old codebase, where we installed Spyne
separately from our LUMC GitHub mirror. This is now also referenced in
the requirements.txt file.

Thanks to Ken Doig for reporting the issue.