Limit input sequence length for description extractor
This is hopefully a temporary measure. At the moment we cannot accurately
predict the running time of the extractor, so we have to aggressively
limit the input based on the worst-case expectation.

As a worst-case scenario, we currently use random input sequences, where
length 1000bp yields about 500ms of running time.

In the future we hope to either:

1. Predict the running time and abort if needed.
2. Keep track of the running time and abort if needed.
3. Run the extractor in a task scheduler with a running time limit.