Create new Redis connection when REDIS_URI changes
When the REDIS_URI configuration setting is changed, the Redis
client should be reconfigured with a new connection pool, just
like we do with the database.

It appears redis-py manages the connection pool by itself and
doesn't expose ways to explicitely destroy it or close all
connections (this is done automatically when all connections
loose scope).

This fix ensures that the unit tests don't accidentally work on
the Redis database configured in MUTALYZER_SETTINGS, which was
quite an unfortunate bug.