Speedup NCBI mapview file import
Instead of querying the existing mappings for overlap and either
updating or inserting depending on the result, we now delete
overlapping mappings first and then only insert.

This is roughly twice as fast. But of course still a horrible
setup compared to some kind of UPSERT functionality which is
unfortunately missing in current PostgreSQL.