Update description-extractor to 2.3.2
Mutalyzer used the `reverse_complement` function from the description
extractor, which was a unicode string-only implementation.

With this update, that implementation was removed and the BioPython
reverse complement function used instead in the description
extractor. Unfortunately this cannot work with unicode strings, so we
now use our own `reverse_complement` which works on unicode strings.