Fix reference file loading by url
For some reason there is no `Content-Type` key in the object returned by
`urllib2.urlopen().info()`. However, the `gettype()` method seems to
work.

I have no idea if this code ever worked, and if so, why. Also, the check
for the `text/plain` content type is pretty bogus in my opinion. A
server can say whatever it likes and valid GenBank files may well be
served with different content types.