Revert adding Reference.version column
While working on this, I came to the conclusion it's not a good idea to
split accession and version. It introduces a lot of complexity for
little benefit.

In general, Mutalyzer always sees 'accession.version' as the identifier
of the reference and because we always want exact identifiers, there is
little need for accession numbers without version.

The most obvious use case I see for a split is that we can easily query
available references with a certain accession, not taking version into
account, as a way to inform the user when a specific reference
identifier was not found. But I guess we better have this use case as
the exception, and make our life easier for the rest.

So I guess I'm aborting this for now. Addition of the `version` column
has already landed in the master branch, but this is easy to roll
back. The original column has not yet been touched in master.